Avoid unnecessary floating-point promotions.

+++ src/GTP.cpp
@@ -61,7 +61,7 @@ int cfg_resignpct;
 int cfg_noise;
 int cfg_random_cnt;
 int cfg_random_min_visits;
-float cfg_random_temp;
+double cfg_random_temp;
 std::uint64_t cfg_rng_seed;
 bool cfg_dumbpass;
 #ifdef USE_OPENCL
@@ -72,9 +72,9 @@ bool cfg_tune_only;
 precision_t cfg_precision;
 #endif
 #endif
-float cfg_puct;
-float cfg_logpuct;
-float cfg_logconst;
+double cfg_puct;
+double cfg_logpuct;
+double cfg_logconst;
 float cfg_softmax_temp;
 float cfg_fpu_reduction;
 float cfg_fpu_root_reduction;
@@ -135,9 +135,9 @@ void GTP::setup_default_parameters() {
     cfg_precision = precision_t::AUTO;
 #endif
 #endif
-    cfg_puct = 0.5f;
-    cfg_logpuct = 0.015f;
-    cfg_logconst = 1.7f;
+    cfg_puct = 0.5;
+    cfg_logpuct = 0.015;
+    cfg_logconst = 1.7;
     cfg_softmax_temp = 1.0f;
     cfg_fpu_reduction = 0.25f;
     // see UCTSearch::should_resign
@@ -146,7 +146,7 @@ void GTP::setup_default_parameters() {
     cfg_fpu_root_reduction = cfg_fpu_reduction;
     cfg_random_cnt = 0;
     cfg_random_min_visits = 1;
-    cfg_random_temp = 1.0f;
+    cfg_random_temp = 1.0;
     cfg_dumbpass = false;
     cfg_logfile_handle = nullptr;
     cfg_quiet = false;
